Enum kvm_ioctls::VcpuExit

pub enum VcpuExit<'a> {

    IoOut(u16, &'a [u8]),
    IoIn(u16, &'a mut [u8]),
    MmioRead(u64, &'a mut [u8]),
    MmioWrite(u64, &'a [u8]),
    Unknown,
    Exception,
    Hypercall,
    Debug(kvm_debug_exit_arch),
    Hlt,
    IrqWindowOpen,
    Shutdown,
    FailEntry,
    Intr,
    SetTpr,
    TprAccess,
    S390Sieic,
    S390Reset,
    Dcr,
    Nmi,
    InternalError,
    Osi,
    PaprHcall,
    S390Ucontrol,
    Watchdog,
    S390Tsch,
    Epr,
    SystemEvent(u32, u64),
    S390Stsi,
    IoapicEoi(u8),
    Hyperv,

}

Reasons for vCPU exits.

The exit reasons are mapped to the KVM_EXIT_* defines in the Linux KVM header.

Variants

IoOut(u16, &'a [u8])

Tuple Fields

0: u16

1: &'a [u8]

An out port instruction was run on the given port with the given data.

IoIn(u16, &'a mut [u8])

Tuple Fields

0: u16

1: &'a mut [u8]

An in port instruction was run on the given port.

The given slice should be filled in before run() is called again.

MmioRead(u64, &'a mut [u8])

Tuple Fields

0: u64

1: &'a mut [u8]

A read instruction was run against the given MMIO address.

The given slice should be filled in before run() is called again.

MmioWrite(u64, &'a [u8])

Tuple Fields

0: u64

1: &'a [u8]

A write instruction was run against the given MMIO address with the given data.

Unknown

Corresponds to KVM_EXIT_UNKNOWN.

Exception

Corresponds to KVM_EXIT_EXCEPTION.

Hypercall

Corresponds to KVM_EXIT_HYPERCALL.

Debug(kvm_debug_exit_arch)

Tuple Fields

0: kvm_debug_exit_arch

Corresponds to KVM_EXIT_DEBUG.

Provides architecture specific information for the debug event.

Hlt

Corresponds to KVM_EXIT_HLT.

IrqWindowOpen

Corresponds to KVM_EXIT_IRQ_WINDOW_OPEN.

Shutdown

Corresponds to KVM_EXIT_SHUTDOWN.

FailEntry

Corresponds to KVM_EXIT_FAIL_ENTRY.

Intr

Corresponds to KVM_EXIT_INTR.

SetTpr

Corresponds to KVM_EXIT_SET_TPR.

TprAccess

Corresponds to KVM_EXIT_TPR_ACCESS.

S390Sieic

Corresponds to KVM_EXIT_S390_SIEIC.

S390Reset

Corresponds to KVM_EXIT_S390_RESET.

Dcr

Corresponds to KVM_EXIT_DCR.

Nmi

Corresponds to KVM_EXIT_NMI.

InternalError

Corresponds to KVM_EXIT_INTERNAL_ERROR.

Osi

Corresponds to KVM_EXIT_OSI.

PaprHcall

Corresponds to KVM_EXIT_PAPR_HCALL.

S390Ucontrol

Corresponds to KVM_EXIT_S390_UCONTROL.

Watchdog

Corresponds to KVM_EXIT_WATCHDOG.

S390Tsch

Corresponds to KVM_EXIT_S390_TSCH.

Epr

Corresponds to KVM_EXIT_EPR.

SystemEvent(u32, u64)

Tuple Fields

0: u32

1: u64

Corresponds to KVM_EXIT_SYSTEM_EVENT.

S390Stsi

Corresponds to KVM_EXIT_S390_STSI.

IoapicEoi(u8)

Tuple Fields

0: u8

Corresponds to KVM_EXIT_IOAPIC_EOI.

Hyperv

Corresponds to KVM_EXIT_HYPERV.